requirements.txt
colorama==0.4.6
docx2pdf==0.1.8
lxml==5.4.0
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
pywin32==310
six==1.17.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
